[IMP] expense : Remove expense warning on accounting dashboard
This commit remove the warnings we have on the accounting dashboard,
as the journals entries are already included in the vendor bills
warnings.

task:3877069
