+++ addons/account/tests/test_account_journal_dashboard.py
@@ -2,33 +2,19 @@ from dateutil.relativedelta import relativedelta
 from freezegun import freeze_time
 
 from odoo import Command
-from odoo.addons.account.tests.common import AccountTestInvoicingCommon
+from odoo.addons.account.tests.test_account_journal_dashboard_common import TestAccountJournalDashboardCommon
 from odoo.tests import tagged
 from odoo.tools.misc import format_amount
 
 @tagged('post_install', '-at_install')
-class TestAccountJournalDashboard(AccountTestInvoicingCommon):
-
-    @classmethod
-    def setUpClass(cls):
-        super().setUpClass()
-        cls.other_currency = cls.setup_other_currency('EUR')
-
-    def assertDashboardPurchaseSaleData(self, journal, number_draft, sum_draft, number_waiting, sum_waiting, number_late, sum_late, currency, **kwargs):
-        expected_values = {
-            'number_draft': number_draft,
-            'sum_draft': format_amount(self.env, sum_draft, currency),
-            'number_waiting': number_waiting,
-            'sum_waiting': format_amount(self.env, sum_waiting, currency),
-            'number_late': number_late,
-            'sum_late': format_amount(self.env, sum_late, currency),
-            **kwargs
-        }
-        dashboard_data = journal._get_journal_dashboard_data_batched()[journal.id]
-        self.assertDictEqual({**dashboard_data, **expected_values}, dashboard_data)
+class TestAccountJournalDashboard(TestAccountJournalDashboardCommon):
 
     @freeze_time("2019-01-22")
     def test_customer_invoice_dashboard(self):
+        # This test is defined in the account_3way_match module with different values, so we skip it when the module is installed
+        if self.env['ir.module.module'].search([('name', '=', 'account_3way_match')]).state == 'installed':
+            self.skipTest("This test won't work if account_3way_match is installed")
+
         journal = self.company_data['default_journal_sale']
 
         invoice = self.env['account.move'].create({
@@ -141,6 +127,10 @@ class TestAccountJournalDashboard(AccountTestInvoicingCommon):
             4) Journal in company currency, bills in company currency -> dashboard data should be displayed in company currency
             5) Journal in company currency, bills in foreign currency -> dashboard data should be displayed in company currency
         """
+        # This test is defined in the account_3way_match module with different values, so we skip it when the module is installed
+        if self.env['ir.module.module'].search([('name', '=', 'account_3way_match')]).state == 'installed':
+            self.skipTest("This test won't work if account_3way_match is installed")
+
         foreign_currency = self.other_currency
         company_currency = self.company_data['currency']
 
@@ -174,6 +164,10 @@ class TestAccountJournalDashboard(AccountTestInvoicingCommon):
                 self.assertDashboardPurchaseSaleData(purchase_journal, *expected_vals)
 
     def test_sale_purchase_journal_for_multi_currency_sale(self):
+        # This test is defined in the account_3way_match module with different values, so we skip it when the module is installed
+        if self.env['ir.module.module'].search([('name', '=', 'account_3way_match')]).state == 'installed':
+            self.skipTest("This test won't work if account_3way_match is installed")
+
         currency = self.other_currency
         company_currency = self.company_data['currency']
 
@@ -210,83 +204,14 @@ class TestAccountJournalDashboard(AccountTestInvoicingCommon):
 
     @freeze_time("2023-03-15")
     def test_purchase_journal_numbers_and_sums(self):
+        # This test is defined in the account_3way_match module with different values, so we skip it when the module is installed
+        if self.env['ir.module.module'].search([('name', '=', 'account_3way_match')]).state == 'installed':
+            self.skipTest("This test won't work if account_3way_match is installed")
+
         company_currency = self.company_data['currency']
         journal = self.company_data['default_journal_purchase']
 
-        #Setup multiple payments term
-        twentyfive_now_term = self.env['account.payment.term'].create({
-            'name': '25% now, rest in 30 days',
-            'note': 'Pay 25% on invoice date and 75% 30 days later',
-            'line_ids': [
-                (0, 0, {
-                    'value': 'percent',
-                    'value_amount': 25.00,
-                    'delay_type': 'days_after',
-                    'nb_days': 0,
-                }),
-                (0, 0, {
-                    'value': 'percent',
-                    'value_amount': 75.00,
-                    'delay_type': 'days_after',
-                    'nb_days': 30,
-                }),
-            ],
-        })
-
-        self.env['account.move'].create({
-            'move_type': 'in_invoice',
-            'journal_id': journal.id,
-            'partner_id': self.partner_a.id,
-            'invoice_date': '2023-04-01',
-            'date': '2023-03-15',
-            'invoice_payment_term_id': twentyfive_now_term.id,
-            'invoice_line_ids': [(0, 0, {
-                'product_id': self.product_a.id,
-                'quantity': 1,
-                'name': 'product test 1',
-                'price_unit': 4000,
-                'tax_ids': [],
-            })]
-        }).action_post()
-        # This bill has two amls of 10$. Both are waiting for payment and due in 16 and 46 days.
-        # number_waiting += 2, sum_waiting += -4000$, number_late += 0, sum_late += 0$
-
-        self.env['account.move'].create({
-            'move_type': 'in_invoice',
-            'journal_id': journal.id,
-            'partner_id': self.partner_a.id,
-            'invoice_date': '2023-03-01',
-            'date': '2023-03-15',
-            'invoice_payment_term_id': twentyfive_now_term.id,
-            'invoice_line_ids': [(0, 0, {
-                'product_id': self.product_a.id,
-                'quantity': 1,
-                'name': 'product test 1',
-                'price_unit': 400,
-                'tax_ids': [],
-            })]
-        }).action_post()
-        # This bill has two amls of 100$. One which is late and due 14 days prior and one which is waiting for payment and due in 15 days.
-        # number_waiting += 2, sum_waiting += -400$, number_late += 1, sum_late += -100$
-
-        self.env['account.move'].create({
-            'move_type': 'in_invoice',
-            'journal_id': journal.id,
-            'partner_id': self.partner_a.id,
-            'invoice_date': '2023-02-01',
-            'date': '2023-03-15',
-            'invoice_payment_term_id': twentyfive_now_term.id,
-            'invoice_line_ids': [(0, 0, {
-                'product_id': self.product_a.id,
-                'quantity': 1,
-                'name': 'product test 1',
-                'price_unit': 40,
-                'tax_ids': [],
-            })]
-        }).action_post()
-        # This bill has two amls of 1000$. Both of them are late and due 45 and 15 days prior.
-        # number_waiting += 2, sum_waiting += -40$, number_late += 2, sum_late += -40$
-
+        self._create_test_vendor_bills(journal)
         dashboard_data = journal._get_journal_dashboard_data_batched()[journal.id]
         # Expected behavior is to have six amls waiting for payment for a total amount of 4440$
         # three of which would be late for a total amount of 140$
