+++ addons/account/views/account_journal_dashboard_view.xml
@@ -345,16 +345,12 @@
                             </div>
                             <div class="row" t-if="dashboard.number_waiting">
                                 <div class="col overflow-hidden text-start">
-                                    <span t-if="journal_type == 'sale'">
-                                        <a type="object" name="open_action" context="{'search_default_open': '1', 'search_default_posted': '1', 'journal_type': type}">
-                                            <t t-out="dashboard.number_waiting"/> Unpaid
-                                        </a>
-                                    </span>
-                                    <span t-if="journal_type == 'purchase'">
-                                        <a type="object" name="open_action" context="{'search_default_open': '1', 'search_default_posted': '1', 'journal_type': type}">
-                                            <t t-out="dashboard.number_waiting"/> To Pay
-                                        </a>
-                                    </span>
+                                    <a t-if="journal_type == 'sale'" type="object" name="open_action" context="{'search_default_open': '1', 'search_default_posted': '1', 'journal_type': type}">
+                                        <span><t t-out="dashboard.number_waiting"/> Unpaid</span>
+                                    </a>
+                                    <a t-if="journal_type == 'purchase'" type="object" name="open_action" context="{'search_default_open': '1', 'search_default_posted': '1', 'journal_type': type}">
+                                        <span title="Bills to Pay" id="account_dashboard_bills_to_pay"><t t-out="dashboard.number_waiting"/> To Pay</span>
+                                    </a>
                                 </div>
                                 <div class="col-auto text-end">
                                     <span class="o_kanban_monetary"><t t-out="dashboard.sum_waiting"/></span>
@@ -362,16 +358,16 @@
                             </div>
                             <div class="row" t-if="dashboard.number_late">
                                 <div class="col overflow-hidden text-start">
-                                    <span t-if="journal_type == 'sale'">
-                                        <a type="object" name="open_action" context="{'search_default_late': '1', 'search_default_posted': '1', 'journal_type': type}">
+                                    <a t-if="journal_type == 'sale'" type="object" name="open_action" context="{'search_default_late': '1', 'search_default_posted': '1', 'journal_type': type}">
+                                        <span title="Invoices late" id="account_dashboard_invoices_late">
                                             <t t-out="dashboard.number_late"/> Late
-                                        </a>
-                                    </span>
-                                    <span t-if="journal_type == 'purchase'">
-                                        <a type="object" name="open_action" context="{'search_default_late': '1', 'search_default_posted': '1', 'journal_type': type}">
+                                        </span>
+                                    </a>
+                                    <a t-if="journal_type == 'purchase'" type="object" name="open_action" context="{'search_default_late': '1', 'search_default_posted': '1', 'journal_type': type}">
+                                        <span title="Bills Late" id="account_dashboard_bills_late">
                                             <t t-out="dashboard.number_late"/> Late
-                                        </a>
-                                    </span>
+                                        </span>
+                                    </a>
                                 </div>
                                 <div class="col-auto text-end">
                                     <span class="o_kanban_monetary"><t t-out="dashboard.sum_late"/></span>
