+++ addons/hr_expense/models/__init__.py
@@ -12,7 +12,6 @@ from . import ir_attachment
 from . import product_product
 from . import product_template
 from . import res_config_settings
-from . import account_journal_dashboard
 from . import res_company
 from . import analytic
 from . import ir_actions_report
