+++ addons/hr_expense/tests/__init__.py
@@ -1,7 +1,6 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import common
-from . import test_account_journal_dashboard
 from . import test_expenses
 from . import test_expenses_access_rights
 from . import test_expenses_mail_import
